# docs-site/openapi.yaml
openapi: 3.0.3
info:
  title: SwipeSavvy Backend API
  description: |
    Multi-service backend API for SwipeSavvy Mobile Wallet — a fintech platform providing
    card issuance (FIS), payments (Authorize.Net), KYC, rewards, real-time chat, and admin management.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via `POST /api/v1/auth/login` + `POST /api/v1/auth/verify-login-otp`.

    | Token | Lifetime | Purpose |
    |-------|----------|---------|
    | Access Token | 24 hours | Authorize API requests |
    | Refresh Token | 30 days | Obtain new access tokens |

    ## Environments
    Use the server selector above to switch between Production, Staging, and Development.
  version: 1.0.0
  contact:
    name: SwipeSavvy Engineering
    email: engineering@swipesavvy.com
  license:
    name: Proprietary
servers:
  - url: http://54.224.8.14:8000
    description: Production (AWS EC2)
  - url: https://api.swipesavvy.com
    description: Production (Domain)
  - url: http://localhost:8000
    description: Development (Local)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login flow
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT token from admin login
    WebhookSignature:
      type: apiKey
      in: header
      name: X-FIS-Signature
      description: FIS webhook signature verification
...
paths:
...
  /charities:
    get:
      tags: [Admin Charities]
      operationId: listCharities
      summary: List charities
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, pending, rejected]
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: Charity list
    post:
      tags: [Admin Charities]
      operationId: createCharity
      summary: Create a charity
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category]
              properties:
                name:
                  type: string
                description:
                  type: string
                category:
                  type: string
                website:
                  type: string
                ein:
                  type: string
                logo_url:
                  type: string
      responses:
        '200':
          description: Charity created
  /charities/{charity_id}:
    get:
      tags: [Admin Charities]
      operationId: getCharity
      summary: Get charity details
      parameters:
        - in: path
          name: charity_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Charity details
    put:
      tags: [Admin Charities]
      operationId: updateCharity
      summary: Update charity
      security:
        - AdminAuth: []
      parameters:
        - in: path
          name: charity_id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                category:
                  type: string
      responses:
        '200':
          description: Charity updated
    delete:
      tags: [Admin Charities]
      operationId: deleteCharity
      summary: Delete charity
      security:
        - AdminAuth: []
      parameters:
        - in: path
          name: charity_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Charity deleted
  /charities/{charity_id}/approve:
    post:
      tags: [Admin Charities]
      operationId: approveCharity
      summary: Approve a charity
      security:
        - AdminAuth: []
      parameters:
        - in: path
          name: charity_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Charity approved
  /charities/{charity_id}/reject:
    post:
      tags: [Admin Charities]
      operationId: rejectCharity
      summary: Reject a charity
      security:
        - AdminAuth: []
      parameters:
        - in: path
          name: charity_id
          required: true
          schema:
            type: string
        - in: query
          name: notes
          schema:
            type: string
      responses:
        '200':
          description: Charity rejected

  # ═══════════════════════════════════════════
  # ADMIN CHAT DASHBOARD (11 endpoints)
  # ═══════════════════════════════════════════
...